axiologic-pla/dev#639 axiologic-pla/dev#640 Update product handling and optimize request handling
This commit updates the product audit function to be asynchronous in the ProductController. It also adds a new product and implements additional tests in the EpiSORClient test suite. Moreover, a common function for processing request parameters has been created to reduce duplication in the EpiSORIntegrationClient. GTIN_SSI function signature in the Product model is corrected as well.