axiologic-pla/dev#640 Refactor batch and product creation in factories
Moved the getEventRecorderInstance method calls inside the factories responsible for creating batch and product instances. This was done to optimize the controllers and to arrange the code more logically. Now, when a batch or a product is created or looked up, it is already given an event recorder instance within the respective factory, simplifying the process in the controller file.