Add authorization middleware for Integration APIs
Implemented a new authorization middleware for the Integration APIs in the middlewares.js file in the utils directory of gtin-resolver. This will enhance resource protection by requiring user permissions checked through their apiKey scopes. Also, minor modification was applied to the messageHeaderSchema.js where the trailing comma on the "messageDateTime" was removed.