Revise request limiter configuration in several files
The configuration for the request limiter was renamed and its capacity values were reduced in several files: "apihub.json", "apihub.json.template", and "middlewares.js". The "throttlerConfig" was renamed to "requestLimiterConfig" and both "metadataCapacity" and "epiCapacity" were reduced from 30 and 10, respectively, to 5. This will control the rate of requests more strictly to prevent overloading.