Improve error handling in data migration
A separation of concerns was introduced in dataMigration.js error handling. The catch blocks have been improved to deal with two distinct sections: migration failure and not being able to mark migration as done. This will allow for more precise error messages and better diagnosis of any issues in the future.